working on generating 500 fake users for database
