+++ data/seeds/001_users.js
@@ -1,22 +1,7 @@
-var faker = require('faker');
-faker.seed(1);
-
-let seeds = [];
-let typearray = ['User', 'Mediator', 'Admin'];
-for (i = 0; i < 100; i++) {
-  let rN = Math.floor(Math.random() * typearray.length);
-  seeds[i] = {
-    username: faker.name.findName(),
-    type: typearray[rN],
-    email: faker.internet.email(),
-    nickname: faker.name.lastName(),
-  };
-};
-
-console.log(seeds);
+const data = require('../fakeUserData');
 
 exports.seed = function(knex) {
-      return knex('users').insert(seeds);
+      return knex('users').insert(data.seeds1);
 };
 
 // {username: 'user1', type: 'user', email: 'user1@email.com', nickname: 'User'},
