+++ data/seeds/002_users.js
@@ -1,0 +1,5 @@
+const data = require('../fakeUserData');
+
+exports.seed = function(knex) {
+  return knex('users').insert(data.seeds2);
+};
