+++ data/seeds/004_users.js
@@ -1,0 +1,6 @@
+const data = require('../fakeUserData');
+
+exports.seed = function(knex) {
+  return knex('users').insert(data.seeds4);
+};
+
